requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
numpy
scipy
scikit-learn
openai
python-dotenv
